agents.py
# ==============================
# Importing Libraries
# ==============================
import os
from dotenv import load_dotenv

# 🔥 Disable OpenAI fallback (IMPORTANT)

# Load environment variables
load_dotenv()

# CrewAI
from crewai import Agent



# Tools
from tools import FinancialDocumentTool

from crewai import Agent, LLM
import os



llm = LLM(
    model="openai/gpt-4o-mini",   
    api_key=os.getenv("OPENAI_API_KEY")
)



# Financial Analyst Agent

financial_analyst = Agent(
    role="Senior Financial Analyst",
    goal=(
        "Analyze the uploaded financial document using the available tool "
        "and provide accurate, data-driven insights for: {query}"
    ),
    verbose=True,
    memory=True,
    backstory=(
        "You are an experienced financial analyst with expertise in analyzing "
        "financial statements, identifying trends, and extracting key metrics "
        "such as revenue, profit, and expenses. You rely strictly on actual "
        "data from the document and avoid speculation."
    ),
    tools=[FinancialDocumentTool()],
    llm=llm,
    allow_delegation=True
)


# Document Verifier Agent

verifier = Agent(
    role="Financial Document Verifier",
    goal=(
        "Verify whether the uploaded file is a valid financial document by "
        "checking for financial terms, structured data, and report patterns."
    ),
    verbose=True,
    memory=True,
    backstory=(
        "You specialize in validating financial documents such as balance sheets, "
        "income statements, and annual reports. You carefully analyze document "
        "content before making a decision."
    ),
    llm=llm,
    allow_delegation=True
)



# Investment Advisor Agent

investment_advisor = Agent(
    role="Investment Advisor",
    goal=(
        "Provide realistic and suitable investment recommendations based on "
        "financial data extracted from the document."
    ),
    verbose=True,
    backstory=(
        "You are a certified financial advisor with strong experience in "
        "portfolio management, risk-return analysis, and market trends. "
        "You provide practical and data-backed investment advice."
    ),
    llm=llm,
    allow_delegation=False
)



# Risk Assessor Agent
risk_assessor = Agent(
    role="Risk Assessment Expert",
    goal=(
        "Identify financial, operational, and market risks based on the "
        "analyzed financial document."
    ),
    verbose=True,
    backstory=(
        "You are an expert in financial risk management with experience in "
        "identifying risk factors such as debt levels, cash flow issues, "
        "market volatility, and operational inefficiencies."
    ),
    llm=llm,
    allow_delegation=False
)