requirements.txt
click
python-dotenv

fastapi
uvicorn

crewai
crewai-tools

langchain
langchain_community
langchain_openai
openai

pypdf

numpy
pandas

pydantic

sqlalchemy

